Finished major stages
Basically we have the basic css, mark up, and code. At the moment we are able to successfully do http requests and get the weather from a city and display a gif describing that weather. However, we will continue to work on how we should handle when there's an error in rendering the weather data, converting units from metric to imperial, having a loading animation with a gif, and properly generating the markdown with the javascript