Major improvments
Now when the user enters an incorrect city, we show a window with an overlay telling them what's wrong. Our fetch functions when possible will throw a response object, which we use to parse a human-readable message from to give users a reason why things went wrong. Instead of hard-coding in imperial or metric units, or percent symbols, we use classes and css to do that for us. Our file module.js now has the corresponding classes that helps with that. All in all, these may look like small improvements, but I think these changes were better for cleaner code that's easier to maintain and modify.