Dark mode add and more code separation
We've successfully implemented a dark mode, which just changes the background image and changes some of colors. More importantly, we've added rendering.js, which is the file we use to put all of our rendering code in that visually changes the screen. I believed that this would better organize the code and at least make it easier to look at.