Added some responsiveness to it.
Add a little responsiveness to the nav in the header, weather-data-grid, and the table. Next time, it would definitely be a better idea to follow a 'mobile first' design if I was wanting to make it responsive. I'll try to do that next time when I'm in the html/css phase of the project.